backend/proactive/requirements.txt
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=4.25.0
firebase-admin>=6.2.0
httpx>=0.27.0
